+++ frontend/src/api/community/article/index.ts
@@ -1,29 +1,12 @@
-import { AxiosError, type AxiosResponse } from 'axios';
+import { type AxiosError, type AxiosResponse } from 'axios';
 import { useMutation, useQuery } from 'react-query';
 
 import { checkType, isNull } from '@utils';
 
-import type { ArticleId, CommunityArticle, Page, Size, StudyId } from '@custom-types';
+import type { ArticleId, CommunityArticle, StudyId } from '@custom-types';
 
 import axiosInstance from '@api/axiosInstance';
-import { checkCommunityArticle, checkCommunityArticles } from '@api/community/typeChecker';
-
-export type ApiCommunityArticles = {
-  get: {
-    responseData: {
-      articles: Array<Omit<CommunityArticle, 'content'>>;
-      currentPage: number;
-      lastPage: number;
-      totalCount: number;
-    };
-    params: {
-      studyId: StudyId;
-      page?: Page;
-      size?: Size;
-    };
-    variables: ApiCommunityArticles['get']['params'];
-  };
-};
+import { checkCommunityArticle } from '@api/community/article/typeChecker';
 
 export type ApiCommunityArticle = {
   get: {
@@ -56,23 +39,6 @@ export type ApiCommunityArticle = {
     };
     variables: ApiCommunityArticle['delete']['params'];
   };
-};
-
-const getCommunityArticles = async ({ studyId, page = 1, size = 8 }: ApiCommunityArticles['get']['variables']) => {
-  // 서버쪽에서는 page를 0번부터 계산하기 때문에 page - 1을 해줘야 한다
-  const response = await axiosInstance.get<ApiCommunityArticles['get']['responseData']>(
-    `/api/studies/${studyId}/community/articles?page=${page - 1}&size=${size}`,
-  );
-
-  return checkCommunityArticles(response.data);
-};
-
-// articles
-export const useGetCommunityArticles = ({ studyId, page }: ApiCommunityArticles['get']['variables']) => {
-  return useQuery<ApiCommunityArticles['get']['responseData'], AxiosError>(
-    ['get-community-articles', studyId, page],
-    () => getCommunityArticles({ studyId, page }),
-  );
 };
 
 // get
