+++ frontend/src/api/community/articles/typeChecker.ts
@@ -4,28 +4,8 @@ import { arrayOfAll, checkType, hasOwnProperties, isArray, isDateYMD, isNumber, 
 
 import type { CommunityArticle } from '@custom-types';
 
-import { type ApiCommunityArticle, type ApiCommunityArticles } from '@api/community';
+import { type ApiCommunityArticles } from '@api/community/articles';
 import { checkMember } from '@api/member/typeChecker';
-
-type CommunityArticleKeys = keyof ApiCommunityArticle['get']['responseData'];
-
-const arrayOfAllCommunityArticleKeys = arrayOfAll<CommunityArticleKeys>();
-
-export const checkCommunityArticle = (data: unknown): ApiCommunityArticle['get']['responseData'] => {
-  if (!isObject(data)) throw new AxiosError(`CommunityArticle does not have correct type: object`);
-
-  const keys = arrayOfAllCommunityArticleKeys(['id', 'author', 'title', 'content', 'createdDate', 'lastModifiedDate']);
-  if (!hasOwnProperties(data, keys)) throw new AxiosError('CommunityArticle does not have some properties');
-
-  return {
-    id: checkType(data.id, isNumber),
-    author: checkMember(data.author),
-    title: checkType(data.title, isString),
-    content: checkType(data.content, isString),
-    createdDate: checkType(data.createdDate, isDateYMD),
-    lastModifiedDate: checkType(data.lastModifiedDate, isDateYMD),
-  };
-};
 
 type Article = Omit<CommunityArticle, 'content'>;
 type ArticleKeys = keyof Article;
